fix: add .int() to ttl schemas to match API validation

+++ packages/core/src/v3/schemas/schemas.ts
@@ -189,7 +189,7 @@ const taskMetadata = {
   triggerSource: z.string().optional(),
   schedule: ScheduleMetadata.optional(),
   maxDuration: z.number().optional(),
-  ttl: z.string().or(z.number().nonnegative()).optional(),
+  ttl: z.string().or(z.number().nonnegative().int()).optional(),
   payloadSchema: z.unknown().optional(),
 };
